+++ R/isoline.R
@@ -18,6 +18,8 @@
 #' @param consumption_model character, specify the consumption model of the vehicle, see \href{https://developer.here.com/documentation/routing-api/8.16.0/dev_guide/topics/use-cases/consumption-model.html}{consumption model} for more information (\code{default = NULL} a average electric car is set).
 #' @param aggregate boolean, aggregate (with function \code{min}) and intersect the isolines from geometry type \code{POLYGON} to geometry type \code{MULTIPOLYGON} (\code{default = TRUE})?
 #' @param url_only boolean, only return the generated URLs (\code{default = FALSE})?
+#' @param type character, 'type' is deprecated, use 'routing_mode' instead.
+#' @param mode character, 'mode' is deprecated, use 'transport_mode' instead.
 #'
 #' @return
 #' An \code{sf} object containing the requested isolines.
@@ -37,7 +39,17 @@ isoline <- function(poi, datetime = Sys.time(), arrival = FALSE,
                     range = seq(5, 30, 5) * 60, range_type = "time",
                     routing_mode = "fast", transport_mode = "car",
                     traffic = TRUE, consumption_model = NULL,
-                    aggregate = TRUE, url_only = FALSE) {
+                    aggregate = TRUE, url_only = FALSE, type, mode) {
+
+  # Deprecated parameters
+  if (!missing("type")) {
+    warning("'type' is deprecated, use 'routing_mode' instead.")
+    routing_mode <- type
+  }
+  if (!missing("mode")) {
+    warning("'mode' is deprecated, use 'transport_mode' instead.")
+    transport_mode <- mode
+  }
 
   # Checks
   .check_points(poi)